Remove redundant self test from Bio.GenBank.Scanner and added test to test_GenBank_unitest.
