+++ Bio/GenBank/Scanner.py
@@ -1721,377 +1721,3 @@ class GenBankScanner(InsdcScanner):
             return
         except StopIteration:
             raise ValueError("Problem in misc lines before sequence")
-
-
-if __name__ == "__main__":
-    from Bio._py3k import StringIO
-
-    gbk_example = \
-        """LOCUS       SCU49845     5028 bp    DNA             PLN       21-JUN-1999
-DEFINITION  Saccharomyces cerevisiae TCP1-beta gene, partial cds, and Axl2p
-            (AXL2) and Rev7p (REV7) genes, complete cds.
-ACCESSION   U49845
-VERSION     U49845.1  GI:1293613
-KEYWORDS    .
-SOURCE      Saccharomyces cerevisiae (baker's yeast)
-  ORGANISM  Saccharomyces cerevisiae
-            Eukaryota; Fungi; Ascomycota; Saccharomycotina; Saccharomycetes;
-            Saccharomycetales; Saccharomycetaceae; Saccharomyces.
-REFERENCE   1  (bases 1 to 5028)
-  AUTHORS   Torpey,L.E., Gibbs,P.E., Nelson,J. and Lawrence,C.W.
-  TITLE     Cloning and sequence of REV7, a gene whose function is required for
-            DNA damage-induced mutagenesis in Saccharomyces cerevisiae
-  JOURNAL   Yeast 10 (11), 1503-1509 (1994)
-  PUBMED    7871890
-REFERENCE   2  (bases 1 to 5028)
-  AUTHORS   Roemer,T., Madden,K., Chang,J. and Snyder,M.
-  TITLE     Selection of axial growth sites in yeast requires Axl2p, a novel
-            plasma membrane glycoprotein
-  JOURNAL   Genes Dev. 10 (7), 777-793 (1996)
-  PUBMED    8846915
-REFERENCE   3  (bases 1 to 5028)
-  AUTHORS   Roemer,T.
-  TITLE     Direct Submission
-  JOURNAL   Submitted (22-FEB-1996) Terry Roemer, Biology, Yale University, New
-            Haven, CT, USA
-FEATURES             Location/Qualifiers
-     source          1..5028
-                     /organism="Saccharomyces cerevisiae"
-                     /db_xref="taxon:4932"
-                     /chromosome="IX"
-                     /map="9"
-     CDS             <1..206
-                     /codon_start=3
-                     /product="TCP1-beta"
-                     /protein_id="AAA98665.1"
-                     /db_xref="GI:1293614"
-                     /translation="SSIYNGISTSGLDLNNGTIADMRQLGIVESYKLKRAVVSSASEA
-                     AEVLLRVDNIIRARPRTANRQHM"
-     gene            687..3158
-                     /gene="AXL2"
-     CDS             687..3158
-                     /gene="AXL2"
-                     /note="plasma membrane glycoprotein"
-                     /codon_start=1
-                     /function="required for axial budding pattern of S.
-                     cerevisiae"
-                     /product="Axl2p"
-                     /protein_id="AAA98666.1"
-                     /db_xref="GI:1293615"
-                     /translation="MTQLQISLLLTATISLLHLVVATPYEAYPIGKQYPPVARVNESF
-                     TFQISNDTYKSSVDKTAQITYNCFDLPSWLSFDSSSRTFSGEPSSDLLSDANTTLYFN
-                     VILEGTDSADSTSLNNTYQFVVTNRPSISLSSDFNLLALLKNYGYTNGKNALKLDPNE
-                     VFNVTFDRSMFTNEESIVSYYGRSQLYNAPLPNWLFFDSGELKFTGTAPVINSAIAPE
-                     TSYSFVIIATDIEGFSAVEVEFELVIGAHQLTTSIQNSLIINVTDTGNVSYDLPLNYV
-                     YLDDDPISSDKLGSINLLDAPDWVALDNATISGSVPDELLGKNSNPANFSVSIYDTYG
-                     DVIYFNFEVVSTTDLFAISSLPNINATRGEWFSYYFLPSQFTDYVNTNVSLEFTNSSQ
-                     DHDWVKFQSSNLTLAGEVPKNFDKLSLGLKANQGSQSQELYFNIIGMDSKITHSNHSA
-                     NATSTRSSHHSTSTSSYTSSTYTAKISSTSAAATSSAPAALPAANKTSSHNKKAVAIA
-                     CGVAIPLGVILVALICFLIFWRRRRENPDDENLPHAISGPDLNNPANKPNQENATPLN
-                     NPFDDDASSYDDTSIARRLAALNTLKLDNHSATESDISSVDEKRDSLSGMNTYNDQFQ
-                     SQSKEELLAKPPVQPPESPFFDPQNRSSSVYMDSEPAVNKSWRYTGNLSPVSDIVRDS
-                     YGSQKTVDTEKLFDLEAPEKEKRTSRDVTMSSLDPWNSNISPSPVRKSVTPSPYNVTK
-                     HRNRHLQNIQDSQSGKNGITPTTMSTSSSDDFVPVKDGENFCWVHSMEPDRRPSKKRL
-                     VDFSNKSNVNVGQVKDIHGRIPEML"
-     gene            complement(3300..4037)
-                     /gene="REV7"
-     CDS             complement(3300..4037)
-                     /gene="REV7"
-                     /codon_start=1
-                     /product="Rev7p"
-                     /protein_id="AAA98667.1"
-                     /db_xref="GI:1293616"
-                     /translation="MNRWVEKWLRVYLKCYINLILFYRNVYPPQSFDYTTYQSFNLPQ
-                     FVPINRHPALIDYIEELILDVLSKLTHVYRFSICIINKKNDLCIEKYVLDFSELQHVD
-                     KDDQIITETEVFDEFRSSLNSLIMHLEKLPKVNDDTITFEAVINAIELELGHKLDRNR
-                     RVDSLEEKAEIERDSNWVKCQEDENLPDNNGFQPPKIKLTSLVGSDVGPLIIHQFSEK
-                     LISGDDKILNGVYSQYEEGESIFGSLF"
-ORIGIN
-        1 gatcctccat atacaacggt atctccacct caggtttaga tctcaacaac ggaaccattg
-       61 ccgacatgag acagttaggt atcgtcgaga gttacaagct aaaacgagca gtagtcagct
-      121 ctgcatctga agccgctgaa gttctactaa gggtggataa catcatccgt gcaagaccaa
-      181 gaaccgccaa tagacaacat atgtaacata tttaggatat acctcgaaaa taataaaccg
-      241 ccacactgtc attattataa ttagaaacag aacgcaaaaa ttatccacta tataattcaa
-      301 agacgcgaaa aaaaaagaac aacgcgtcat agaacttttg gcaattcgcg tcacaaataa
-      361 attttggcaa cttatgtttc ctcttcgagc agtactcgag ccctgtctca agaatgtaat
-      421 aatacccatc gtaggtatgg ttaaagatag catctccaca acctcaaagc tccttgccga
-      481 gagtcgccct cctttgtcga gtaattttca cttttcatat gagaacttat tttcttattc
-      541 tttactctca catcctgtag tgattgacac tgcaacagcc accatcacta gaagaacaga
-      601 acaattactt aatagaaaaa ttatatcttc ctcgaaacga tttcctgctt ccaacatcta
-      661 cgtatatcaa gaagcattca cttaccatga cacagcttca gatttcatta ttgctgacag
-      721 ctactatatc actactccat ctagtagtgg ccacgcccta tgaggcatat cctatcggaa
-      781 aacaataccc cccagtggca agagtcaatg aatcgtttac atttcaaatt tccaatgata
-      841 cctataaatc gtctgtagac aagacagctc aaataacata caattgcttc gacttaccga
-      901 gctggctttc gtttgactct agttctagaa cgttctcagg tgaaccttct tctgacttac
-      961 tatctgatgc gaacaccacg ttgtatttca atgtaatact cgagggtacg gactctgccg
-     1021 acagcacgtc tttgaacaat acataccaat ttgttgttac aaaccgtcca tccatctcgc
-     1081 tatcgtcaga tttcaatcta ttggcgttgt taaaaaacta tggttatact aacggcaaaa
-     1141 acgctctgaa actagatcct aatgaagtct tcaacgtgac ttttgaccgt tcaatgttca
-     1201 ctaacgaaga atccattgtg tcgtattacg gacgttctca gttgtataat gcgccgttac
-     1261 ccaattggct gttcttcgat tctggcgagt tgaagtttac tgggacggca ccggtgataa
-     1321 actcggcgat tgctccagaa acaagctaca gttttgtcat catcgctaca gacattgaag
-     1381 gattttctgc cgttgaggta gaattcgaat tagtcatcgg ggctcaccag ttaactacct
-     1441 ctattcaaaa tagtttgata atcaacgtta ctgacacagg taacgtttca tatgacttac
-     1501 ctctaaacta tgtttatctc gatgacgatc ctatttcttc tgataaattg ggttctataa
-     1561 acttattgga tgctccagac tgggtggcat tagataatgc taccatttcc gggtctgtcc
-     1621 cagatgaatt actcggtaag aactccaatc ctgccaattt ttctgtgtcc atttatgata
-     1681 cttatggtga tgtgatttat ttcaacttcg aagttgtctc cacaacggat ttgtttgcca
-     1741 ttagttctct tcccaatatt aacgctacaa ggggtgaatg gttctcctac tattttttgc
-     1801 cttctcagtt tacagactac gtgaatacaa acgtttcatt agagtttact aattcaagcc
-     1861 aagaccatga ctgggtgaaa ttccaatcat ctaatttaac attagctgga gaagtgccca
-     1921 agaatttcga caagctttca ttaggtttga aagcgaacca aggttcacaa tctcaagagc
-     1981 tatattttaa catcattggc atggattcaa agataactca ctcaaaccac agtgcgaatg
-     2041 caacgtccac aagaagttct caccactcca cctcaacaag ttcttacaca tcttctactt
-     2101 acactgcaaa aatttcttct acctccgctg ctgctacttc ttctgctcca gcagcgctgc
-     2161 cagcagccaa taaaacttca tctcacaata aaaaagcagt agcaattgcg tgcggtgttg
-     2221 ctatcccatt aggcgttatc ctagtagctc tcatttgctt cctaatattc tggagacgca
-     2281 gaagggaaaa tccagacgat gaaaacttac cgcatgctat tagtggacct gatttgaata
-     2341 atcctgcaaa taaaccaaat caagaaaacg ctacaccttt gaacaacccc tttgatgatg
-     2401 atgcttcctc gtacgatgat acttcaatag caagaagatt ggctgctttg aacactttga
-     2461 aattggataa ccactctgcc actgaatctg atatttccag cgtggatgaa aagagagatt
-     2521 ctctatcagg tatgaataca tacaatgatc agttccaatc ccaaagtaaa gaagaattat
-     2581 tagcaaaacc cccagtacag cctccagaga gcccgttctt tgacccacag aataggtctt
-     2641 cttctgtgta tatggatagt gaaccagcag taaataaatc ctggcgatat actggcaacc
-     2701 tgtcaccagt ctctgatatt gtcagagaca gttacggatc acaaaaaact gttgatacag
-     2761 aaaaactttt cgatttagaa gcaccagaga aggaaaaacg tacgtcaagg gatgtcacta
-     2821 tgtcttcact ggacccttgg aacagcaata ttagcccttc tcccgtaaga aaatcagtaa
-     2881 caccatcacc atataacgta acgaagcatc gtaaccgcca cttacaaaat attcaagact
-     2941 ctcaaagcgg taaaaacgga atcactccca caacaatgtc aacttcatct tctgacgatt
-     3001 ttgttccggt taaagatggt gaaaattttt gctgggtcca tagcatggaa ccagacagaa
-     3061 gaccaagtaa gaaaaggtta gtagattttt caaataagag taatgtcaat gttggtcaag
-     3121 ttaaggacat tcacggacgc atcccagaaa tgctgtgatt atacgcaacg atattttgct
-     3181 taattttatt ttcctgtttt attttttatt agtggtttac agatacccta tattttattt
-     3241 agtttttata cttagagaca tttaatttta attccattct tcaaatttca tttttgcact
-     3301 taaaacaaag atccaaaaat gctctcgccc tcttcatatt gagaatacac tccattcaaa
-     3361 attttgtcgt caccgctgat taatttttca ctaaactgat gaataatcaa aggccccacg
-     3421 tcagaaccga ctaaagaagt gagttttatt ttaggaggtt gaaaaccatt attgtctggt
-     3481 aaattttcat cttcttgaca tttaacccag tttgaatccc tttcaatttc tgctttttcc
-     3541 tccaaactat cgaccctcct gtttctgtcc aacttatgtc ctagttccaa ttcgatcgca
-     3601 ttaataactg cttcaaatgt tattgtgtca tcgttgactt taggtaattt ctccaaatgc
-     3661 ataatcaaac tatttaagga agatcggaat tcgtcgaaca cttcagtttc cgtaatgatc
-     3721 tgatcgtctt tatccacatg ttgtaattca ctaaaatcta aaacgtattt ttcaatgcat
-     3781 aaatcgttct ttttattaat aatgcagatg gaaaatctgt aaacgtgcgt taatttagaa
-     3841 agaacatcca gtataagttc ttctatatag tcaattaaag caggatgcct attaatggga
-     3901 acgaactgcg gcaagttgaa tgactggtaa gtagtgtagt cgaatgactg aggtgggtat
-     3961 acatttctat aaaataaaat caaattaatg tagcatttta agtataccct cagccacttc
-     4021 tctacccatc tattcataaa gctgacgcaa cgattactat tttttttttc ttcttggatc
-     4081 tcagtcgtcg caaaaacgta taccttcttt ttccgacctt ttttttagct ttctggaaaa
-     4141 gtttatatta gttaaacagg gtctagtctt agtgtgaaag ctagtggttt cgattgactg
-     4201 atattaagaa agtggaaatt aaattagtag tgtagacgta tatgcatatg tatttctcgc
-     4261 ctgtttatgt ttctacgtac ttttgattta tagcaagggg aaaagaaata catactattt
-     4321 tttggtaaag gtgaaagcat aatgtaaaag ctagaataaa atggacgaaa taaagagagg
-     4381 cttagttcat cttttttcca aaaagcaccc aatgataata actaaaatga aaaggatttg
-     4441 ccatctgtca gcaacatcag ttgtgtgagc aataataaaa tcatcacctc cgttgccttt
-     4501 agcgcgtttg tcgtttgtat cttccgtaat tttagtctta tcaatgggaa tcataaattt
-     4561 tccaatgaat tagcaatttc gtccaattct ttttgagctt cttcatattt gctttggaat
-     4621 tcttcgcact tcttttccca ttcatctctt tcttcttcca aagcaacgat ccttctaccc
-     4681 atttgctcag agttcaaatc ggcctctttc agtttatcca ttgcttcctt cagtttggct
-     4741 tcactgtctt ctagctgttg ttctagatcc tggtttttct tggtgtagtt ctcattatta
-     4801 gatctcaagt tattggagtc ttcagccaat tgctttgtat cagacaattg actctctaac
-     4861 ttctccactt cactgtcgag ttgctcgttt ttagcggaca aagatttaat ctcgttttct
-     4921 ttttcagtgt tagattgctc taattctttg agctgttctc tcagctcctc atatttttct
-     4981 tgccatgact cagattctaa ttttaagcta ttcaatttct ctttgatc
-//"""
-
-    # GenBank format protein (aka GenPept) file from:
-    # http://www.molecularevolution.org/resources/fileformats/
-    gbk_example2 = \
-        """LOCUS       AAD51968                 143 aa            linear   BCT 21-AUG-2001
-DEFINITION  transcriptional regulator RovA [Yersinia enterocolitica].
-ACCESSION   AAD51968
-VERSION     AAD51968.1  GI:5805369
-DBSOURCE    locus AF171097 accession AF171097.1
-KEYWORDS    .
-SOURCE      Yersinia enterocolitica
-  ORGANISM  Yersinia enterocolitica
-            Bacteria; Proteobacteria; Gammaproteobacteria; Enterobacteriales;
-            Enterobacteriaceae; Yersinia.
-REFERENCE   1  (residues 1 to 143)
-  AUTHORS   Revell,P.A. and Miller,V.L.
-  TITLE     A chromosomally encoded regulator is required for expression of the
-            Yersinia enterocolitica inv gene and for virulence
-  JOURNAL   Mol. Microbiol. 35 (3), 677-685 (2000)
-  MEDLINE   20138369
-   PUBMED   10672189
-REFERENCE   2  (residues 1 to 143)
-  AUTHORS   Revell,P.A. and Miller,V.L.
-  TITLE     Direct Submission
-  JOURNAL   Submitted (22-JUL-1999) Molecular Microbiology, Washington
-            University School of Medicine, Campus Box 8230, 660 South Euclid,
-            St. Louis, MO 63110, USA
-COMMENT     Method: conceptual translation.
-FEATURES             Location/Qualifiers
-     source          1..143
-                     /organism="Yersinia enterocolitica"
-                     /mol_type="unassigned DNA"
-                     /strain="JB580v"
-                     /serotype="O:8"
-                     /db_xref="taxon:630"
-     Protein         1..143
-                     /product="transcriptional regulator RovA"
-                     /name="regulates inv expression"
-     CDS             1..143
-                     /gene="rovA"
-                     /coded_by="AF171097.1:380..811"
-                     /note="regulator of virulence"
-                     /transl_table=11
-ORIGIN
-        1 mestlgsdla rlvrvwrali dhrlkplelt qthwvtlhni nrlppeqsqi qlakaigieq
-       61 pslvrtldql eekglitrht candrrakri klteqsspii eqvdgvicst rkeilggisp
-      121 deiellsgli dklerniiql qsk
-//
-"""
-
-    embl_example = """ID   X56734; SV 1; linear; mRNA; STD; PLN; 1859 BP.
-XX
-AC   X56734; S46826;
-XX
-DT   12-SEP-1991 (Rel. 29, Created)
-DT   25-NOV-2005 (Rel. 85, Last updated, Version 11)
-XX
-DE   Trifolium repens mRNA for non-cyanogenic beta-glucosidase
-XX
-KW   beta-glucosidase.
-XX
-OS   Trifolium repens (white clover)
-OC   Eukaryota; Viridiplantae; Streptophyta; Embryophyta; Tracheophyta;
-OC   Spermatophyta; Magnoliophyta; eudicotyledons; core eudicotyledons; rosids;
-OC   eurosids I; Fabales; Fabaceae; Papilionoideae; Trifolieae; Trifolium.
-XX
-RN   [5]
-RP   1-1859
-RX   PUBMED; 1907511.
-RA   Oxtoby E., Dunn M.A., Pancoro A., Hughes M.A.;
-RT   "Nucleotide and derived amino acid sequence of the cyanogenic
-RT   beta-glucosidase (linamarase) from white clover (Trifolium repens L.)";
-RL   Plant Mol. Biol. 17(2):209-219(1991).
-XX
-RN   [6]
-RP   1-1859
-RA   Hughes M.A.;
-RT   ;
-RL   Submitted (19-NOV-1990) to the EMBL/GenBank/DDBJ databases.
-RL   Hughes M.A., University of Newcastle Upon Tyne, Medical School, Newcastle
-RL   Upon Tyne, NE2 4HH, UK
-XX
-FH   Key             Location/Qualifiers
-FH
-FT   source          1..1859
-FT                   /organism="Trifolium repens"
-FT                   /mol_type="mRNA"
-FT                   /clone_lib="lambda gt10"
-FT                   /clone="TRE361"
-FT                   /tissue_type="leaves"
-FT                   /db_xref="taxon:3899"
-FT   CDS             14..1495
-FT                   /product="beta-glucosidase"
-FT                   /EC_number="3.2.1.21"
-FT                   /note="non-cyanogenic"
-FT                   /db_xref="GOA:P26204"
-FT                   /db_xref="InterPro:IPR001360"
-FT                   /db_xref="InterPro:IPR013781"
-FT                   /db_xref="UniProtKB/Swiss-Prot:P26204"
-FT                   /protein_id="CAA40058.1"
-FT                   /translation="MDFIVAIFALFVISSFTITSTNAVEASTLLDIGNLSRSSFPRGFI
-FT                   FGAGSSAYQFEGAVNEGGRGPSIWDTFTHKYPEKIRDGSNADITVDQYHRYKEDVGIMK
-FT                   DQNMDSYRFSISWPRILPKGKLSGGINHEGIKYYNNLINELLANGIQPFVTLFHWDLPQ
-FT                   VLEDEYGGFLNSGVINDFRDYTDLCFKEFGDRVRYWSTLNEPWVFSNSGYALGTNAPGR
-FT                   CSASNVAKPGDSGTGPYIVTHNQILAHAEAVHVYKTKYQAYQKGKIGITLVSNWLMPLD
-FT                   DNSIPDIKAAERSLDFQFGLFMEQLTTGDYSKSMRRIVKNRLPKFSKFESSLVNGSFDF
-FT                   IGINYYSSSYISNAPSHGNAKPSYSTNPMTNISFEKHGIPLGPRAASIWIYVYPYMFIQ
-FT                   EDFEIFCYILKINITILQFSITENGMNEFNDATLPVEEALLNTYRIDYYYRHLYYIRSA
-FT                   IRAGSNVKGFYAWSFLDCNEWFAGFTVRFGLNFVD"
-FT   mRNA            1..1859
-FT                   /experiment="experimental evidence, no additional details
-FT                   recorded"
-XX
-SQ   Sequence 1859 BP; 609 A; 314 C; 355 G; 581 T; 0 other;
-     aaacaaacca aatatggatt ttattgtagc catatttgct ctgtttgtta ttagctcatt        60
-     cacaattact tccacaaatg cagttgaagc ttctactctt cttgacatag gtaacctgag       120
-     tcggagcagt tttcctcgtg gcttcatctt tggtgctgga tcttcagcat accaatttga       180
-     aggtgcagta aacgaaggcg gtagaggacc aagtatttgg gataccttca cccataaata       240
-     tccagaaaaa ataagggatg gaagcaatgc agacatcacg gttgaccaat atcaccgcta       300
-     caaggaagat gttgggatta tgaaggatca aaatatggat tcgtatagat tctcaatctc       360
-     ttggccaaga atactcccaa agggaaagtt gagcggaggc ataaatcacg aaggaatcaa       420
-     atattacaac aaccttatca acgaactatt ggctaacggt atacaaccat ttgtaactct       480
-     ttttcattgg gatcttcccc aagtcttaga agatgagtat ggtggtttct taaactccgg       540
-     tgtaataaat gattttcgag actatacgga tctttgcttc aaggaatttg gagatagagt       600
-     gaggtattgg agtactctaa atgagccatg ggtgtttagc aattctggat atgcactagg       660
-     aacaaatgca ccaggtcgat gttcggcctc caacgtggcc aagcctggtg attctggaac       720
-     aggaccttat atagttacac acaatcaaat tcttgctcat gcagaagctg tacatgtgta       780
-     taagactaaa taccaggcat atcaaaaggg aaagataggc ataacgttgg tatctaactg       840
-     gttaatgcca cttgatgata atagcatacc agatataaag gctgccgaga gatcacttga       900
-     cttccaattt ggattgttta tggaacaatt aacaacagga gattattcta agagcatgcg       960
-     gcgtatagtt aaaaaccgat tacctaagtt ctcaaaattc gaatcaagcc tagtgaatgg      1020
-     ttcatttgat tttattggta taaactatta ctcttctagt tatattagca atgccccttc      1080
-     acatggcaat gccaaaccca gttactcaac aaatcctatg accaatattt catttgaaaa      1140
-     acatgggata cccttaggtc caagggctgc ttcaatttgg atatatgttt atccatatat      1200
-     gtttatccaa gaggacttcg agatcttttg ttacatatta aaaataaata taacaatcct      1260
-     gcaattttca atcactgaaa atggtatgaa tgaattcaac gatgcaacac ttccagtaga      1320
-     agaagctctt ttgaatactt acagaattga ttactattac cgtcacttat actacattcg      1380
-     ttctgcaatc agggctggct caaatgtgaa gggtttttac gcatggtcat ttttggactg      1440
-     taatgaatgg tttgcaggct ttactgttcg ttttggatta aactttgtag attagaaaga      1500
-     tggattaaaa aggtacccta agctttctgc ccaatggtac aagaactttc tcaaaagaaa      1560
-     ctagctagta ttattaaaag aactttgtag tagattacag tacatcgttt gaagttgagt      1620
-     tggtgcacct aattaaataa aagaggttac tcttaacata tttttaggcc attcgttgtg      1680
-     aagttgttag gctgttattt ctattatact atgttgtagt aataagtgca ttgttgtacc      1740
-     agaagctatg atcataacta taggttgatc cttcatgtat cagtttgatg ttgagaatac      1800
-     tttgaattaa aagtcttttt ttattttttt aaaaaaaaaa aaaaaaaaaa aaaaaaaaa       1859
-//
-"""
-
-    print("GenBank CDS Iteration")
-    print("=====================")
-
-    g = GenBankScanner()
-    for record in g.parse_cds_features(StringIO(gbk_example)):
-        print(record)
-
-    g = GenBankScanner()
-    for record in g.parse_cds_features(StringIO(gbk_example2),
-                                       tags2id=('gene', 'locus_tag', 'product')):
-        print(record)
-
-    g = GenBankScanner()
-    for record in g.parse_cds_features(StringIO(gbk_example + "\n" + gbk_example2),
-                                       tags2id=('gene', 'locus_tag', 'product')):
-        print(record)
-
-    print("")
-    print("GenBank Iteration")
-    print("=================")
-    g = GenBankScanner()
-    for record in g.parse_records(StringIO(gbk_example), do_features=False):
-        print("%s %s %s" % (record.id, record.name, record.description))
-        print(record.seq)
-
-    g = GenBankScanner()
-    for record in g.parse_records(StringIO(gbk_example), do_features=True):
-        print("%s %s %s" % (record.id, record.name, record.description))
-        print(record.seq)
-
-    g = GenBankScanner()
-    for record in g.parse_records(StringIO(gbk_example2), do_features=False):
-        print("%s %s %s" % (record.id, record.name, record.description))
-        print(record.seq)
-
-    g = GenBankScanner()
-    for record in g.parse_records(StringIO(gbk_example2), do_features=True):
-        print("%s %s %s" % (record.id, record.name, record.description))
-        print(record.seq)
-
-    print("")
-    print("EMBL CDS Iteration")
-    print("==================")
-
-    e = EmblScanner()
-    for record in e.parse_cds_features(StringIO(embl_example)):
-        print(record)
-
-    print("")
-    print("EMBL Iteration")
-    print("==============")
-    e = EmblScanner()
-    for record in e.parse_records(StringIO(embl_example), do_features=True):
-        print("%s %s %s" % (record.id, record.name, record.description))
-        print(record.seq)
